Actualizo el stock de productos, tras realizar pedido, terminado.

+++ application/controllers/RealizarPedido.php
@@ -11,6 +11,7 @@ class RealizarPedido extends CI_Controller
         $this->load->model('usuario');
         $this->load->model("pedidos");
         $this->load->library("carrito");
+        $this->load->model('productos');
     }
 
     public function index()
@@ -25,20 +26,41 @@ class RealizarPedido extends CI_Controller
         // 4º Obtengo los productos almacenados en el carrito de compra.
         $listaProductos = $this->carrito->dameTodosLosProductos();
 
-
         // Recorro la lista de productos
         foreach ($listaProductos as $libro) {
 
-            $item = array(
+            // Guardo en un array, los stock de los productos ya actualizados.
+            $stocks[] = array(
                 "idProducto" => $libro[0]["idProducto"],
+                "stock" =>  $libro[0]["stock"] - $libro[0]["cantidad"]
+            );
+            // Guardo en un array, todos los productos del carrito de compra.
+            $items[] = array(
                 "precio" => $libro[0]["precio"],
                 "cantidad" =>  $libro[0]["cantidad"],
-                "idPedido" => $idPedido
+                "idPedido" => $idPedido,
+                "idProducto" => $libro[0]["idProducto"]
             );
+        }
+        echo "Items <br>";
 
-            // Inserto los items en la tabla linea de pedido.
-            $this->pedidos->insertarProductosEnLineaDePedido($item);
+        echo "<pre>";
+        print_r($items);
+        echo "</pre>";
+
+        echo "Stocks: <br>";
+
+        echo "<pre>";
+        print_r($stocks);
+        echo "</pre>";
+
+
+        // Inserto los items en la tabla linea de pedido.
+        if ($this->pedidos->insertarProductosEnLineaDePedido($items) and  $this->productos->actualizarStock($stocks)) {
             $this->carrito->destroy();
+            echo "PEDIDO REALIZADO";
+        } else {
+            echo "ERROR";
         }
     }
 }// Final clase
